Feature/31 orders boxes product in boxes

+++ backend/src/main/java/com/project/harmonie_e_commerce/model/CompositeKey/ProductInCartKey.java
@@ -14,7 +14,7 @@ import java.io.Serializable;
 @NoArgsConstructor
 public class ProductInCartKey implements Serializable {
     @Column(name = "cart_id")
-    private Long cartId;
+    private Integer cartId;
     @Column(name = "product_id")
-    private Long productId;
+    private Integer productId;
 }
